Move buy-buttons into collocated folder with store and sticky bar
